refactor: move event emitter api to common
