+++ filenames.gni
@@ -275,8 +275,6 @@ filenames = {
     "shell/browser/api/electron_api_download_item.cc",
     "shell/browser/api/electron_api_download_item.h",
     "shell/browser/api/electron_api_event.cc",
-    "shell/browser/api/electron_api_event_emitter.cc",
-    "shell/browser/api/electron_api_event_emitter.h",
     "shell/browser/api/electron_api_global_shortcut.cc",
     "shell/browser/api/electron_api_global_shortcut.h",
     "shell/browser/api/electron_api_in_app_purchase.cc",
@@ -527,6 +525,8 @@ filenames = {
     "shell/common/api/electron_api_clipboard.h",
     "shell/common/api/electron_api_command_line.cc",
     "shell/common/api/electron_api_environment.cc",
+    "shell/common/api/electron_api_event_emitter.cc",
+    "shell/common/api/electron_api_event_emitter.h",
     "shell/common/api/electron_api_key_weak_map.h",
     "shell/common/api/electron_api_native_image.cc",
     "shell/common/api/electron_api_native_image.h",
