+++ shell/common/api/electron_api_event_emitter.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#include "shell/browser/api/electron_api_event_emitter.h"
+#include "shell/common/api/electron_api_event_emitter.h"
 
 #include "base/bind.h"
 #include "base/callback.h"
@@ -46,4 +46,4 @@ v8::Local<v8::Object> GetEventEmitterPrototype(v8::Isolate* isolate) {
 
 }  // namespace electron
 
-NODE_LINKED_MODULE_CONTEXT_AWARE(electron_browser_event_emitter, Initialize)
+NODE_LINKED_MODULE_CONTEXT_AWARE(electron_common_event_emitter, Initialize)
